feat(app): 应用目录调整和代码模板调整
