+++ cmd/app/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "go-gin-web/internal/app"
+
+func main() {
+	app.Run()
+}
