--- cmd/demo/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "go-gin-web/internal/demo"
-
-func main() {
-	demo.Run()
-}
