+++ internal/app/dto/base.go
