+++ internal/app/helper/close.go
