+++ internal/app/router/enter.go
@@ -3,5 +3,5 @@ package router
 import "github.com/gin-gonic/gin"
 
 func RegisterRouter(routerGroup *gin.RouterGroup) {
-	registerUserRouter(routerGroup)
+	userRouter(routerGroup)
 }
