+++ internal/genCode/config/config.yaml
@@ -1,5 +1,5 @@
 mysql:
-  service: demo
+  service: genCode
   addr: 127.0.0.1:3306
   user: root
   password: 123456
@@ -21,32 +21,32 @@ code_gen:
   mode: api
   module:
     tpl_dir: internal/resource/codeTpl/module
-    root_dir: internal/demo
-    import_dir_prefix: go-gin-web/internal/demo
+    root_dir: internal/app
+    import_dir_prefix: go-gin-web/internal/app
     description: 平台
     api_doc_tag: 平台管理
     api_group: platform
-    api_prefix: /demo/platform/
+    api_prefix: /app/platform/
     package_name: platform
     table_name: platform
   model:
     tpl_dir: internal/resource/codeTpl/model
-    root_dir: internal/demo
-    import_dir_prefix: go-gin-web/internal/demo
+    root_dir: internal/app
+    import_dir_prefix: go-gin-web/internal/app
     description: 平台
     package_name: platform
     table_name: platform_application
   api:
     tpl_dir: internal/resource/codeTpl/api
-    root_dir: internal/demo
-    import_dir_prefix: go-gin-web/internal/demo
-    description: 平台
+    root_dir: internal/app
+    import_dir_prefix: go-gin-web/internal/app
+    description: 平台测试
     package_name: platform
     target_filename: platform.go
     http_method: POST
     function_name: Test2
     api_doc_tag: 平台管理
     api_group: platform
-    api_prefix: /demo/platform/
+    api_prefix: /app/platform/
     api_suffix: test
     routerGroup: platformGroup
