+++ internal/resource/codeTpl/module/router.go.tpl
@@ -7,14 +7,14 @@ import (
 )
 
 // {{.ReceiverTypeName}}Router 初始化{{.Description}}路由信息
-func {{.ReceiverTypeName}}Router(privateRouter *gin.RouterGroup) {
+func {{.ReceiverTypeName}}Router(routerGroup *gin.RouterGroup) {
 	{{.ReceiverTypeName}}Ctr := ctr{{.PackagePascalName}}.New{{.ReceiverTypePascalName}}Ctr()
-	routerGroup := privateRouter.Group("{{.ApiGroup}}")
+	{{.ReceiverTypeName}}Group := routerGroup.Group("{{.ApiGroup}}")
 	{
-		routerGroup.POST("create", {{.ReceiverTypeName}}Ctr.Create)   // 新建{{.Description}}
-		routerGroup.POST("delete", {{.ReceiverTypeName}}Ctr.Delete)   // 删除{{.Description}}
-		routerGroup.POST("update", {{.ReceiverTypeName}}Ctr.Update)   // 更新{{.Description}}
-		routerGroup.GET("detail", {{.ReceiverTypeName}}Ctr.Detail)    // 根据ID获取{{.Description}}
-        routerGroup.GET("pageList", {{.ReceiverTypeName}}Ctr.PageList)  // 获取{{.Description}}列表
+		{{.ReceiverTypeName}}Group.POST("create", {{.ReceiverTypeName}}Ctr.Create)   // 新建{{.Description}}
+		{{.ReceiverTypeName}}Group.POST("delete", {{.ReceiverTypeName}}Ctr.Delete)   // 删除{{.Description}}
+		{{.ReceiverTypeName}}Group.POST("update", {{.ReceiverTypeName}}Ctr.Update)   // 更新{{.Description}}
+		{{.ReceiverTypeName}}Group.GET("detail", {{.ReceiverTypeName}}Ctr.Detail)    // 根据ID获取{{.Description}}
+        {{.ReceiverTypeName}}Group.GET("pageList", {{.ReceiverTypeName}}Ctr.PageList)  // 获取{{.Description}}列表
 	}
 }
